# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAB SCHOOL CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type StaffMemberRole =
  (typeof StaffMemberRole)[keyof typeof StaffMemberRole];

export const StaffMemberRole = {
  admin: "admin",
  manager: "manager",
  assistant: "assistant",
  staff: "staff",
} as const;

export interface StaffMember {
  id: number;
  username: string;
  fullName: string;
  role: StaffMemberRole;
  createdAt: string;
}

export type CreateStaffBodyRole =
  (typeof CreateStaffBodyRole)[keyof typeof CreateStaffBodyRole];

export const CreateStaffBodyRole = {
  admin: "admin",
  manager: "manager",
  assistant: "assistant",
  staff: "staff",
} as const;

export interface CreateStaffBody {
  username: string;
  password: string;
  fullName: string;
  role: CreateStaffBodyRole;
}

export type UpdateStaffBodyRole =
  (typeof UpdateStaffBodyRole)[keyof typeof UpdateStaffBodyRole];

export const UpdateStaffBodyRole = {
  admin: "admin",
  manager: "manager",
  assistant: "assistant",
  staff: "staff",
} as const;

export interface UpdateStaffBody {
  fullName?: string;
  role?: UpdateStaffBodyRole;
  password?: string;
}

export type StudentTrainingType =
  (typeof StudentTrainingType)[keyof typeof StudentTrainingType];

export const StudentTrainingType = {
  online: "online",
  physical: "physical",
} as const;

export type StudentStage = (typeof StudentStage)[keyof typeof StudentStage];

export const StudentStage = {
  new: "new",
  contacted: "contacted",
  interested: "interested",
  no_show: "no_show",
  archived: "archived",
} as const;

export interface Student {
  id: number;
  firstName: string;
  lastName: string;
  phone: string;
  whatsapp: string;
  city: string;
  trainingType: StudentTrainingType;
  housingNeeded: boolean;
  experienceLevel: string;
  /** @nullable */
  note?: string | null;
  stage: StudentStage;
  /** @nullable */
  groupId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateStudentBodyTrainingType =
  (typeof CreateStudentBodyTrainingType)[keyof typeof CreateStudentBodyTrainingType];

export const CreateStudentBodyTrainingType = {
  online: "online",
  physical: "physical",
} as const;

export interface CreateStudentBody {
  firstName: string;
  lastName: string;
  phone: string;
  whatsapp: string;
  city: string;
  trainingType: CreateStudentBodyTrainingType;
  housingNeeded: boolean;
  experienceLevel: string;
  note?: string;
}

export type UpdateStudentBodyTrainingType =
  (typeof UpdateStudentBodyTrainingType)[keyof typeof UpdateStudentBodyTrainingType];

export const UpdateStudentBodyTrainingType = {
  online: "online",
  physical: "physical",
} as const;

export type UpdateStudentBodyStage =
  (typeof UpdateStudentBodyStage)[keyof typeof UpdateStudentBodyStage];

export const UpdateStudentBodyStage = {
  new: "new",
  contacted: "contacted",
  interested: "interested",
  no_show: "no_show",
  archived: "archived",
} as const;

export interface UpdateStudentBody {
  firstName?: string;
  lastName?: string;
  phone?: string;
  whatsapp?: string;
  city?: string;
  trainingType?: UpdateStudentBodyTrainingType;
  housingNeeded?: boolean;
  experienceLevel?: string;
  /** @nullable */
  note?: string | null;
  stage?: UpdateStudentBodyStage;
}

export type UpdateStageBodyStage =
  (typeof UpdateStageBodyStage)[keyof typeof UpdateStageBodyStage];

export const UpdateStageBodyStage = {
  new: "new",
  contacted: "contacted",
  interested: "interested",
  no_show: "no_show",
  archived: "archived",
} as const;

export interface UpdateStageBody {
  stage: UpdateStageBodyStage;
}

export interface AssignGroupBody {
  /** @nullable */
  groupId: number | null;
}

export type GroupTrainingType =
  (typeof GroupTrainingType)[keyof typeof GroupTrainingType];

export const GroupTrainingType = {
  online: "online",
  physical: "physical",
} as const;

export type GroupStatus = (typeof GroupStatus)[keyof typeof GroupStatus];

export const GroupStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface Group {
  id: number;
  name: string;
  startDate: string;
  trainingType: GroupTrainingType;
  capacity: number;
  /** @nullable */
  notes?: string | null;
  status: GroupStatus;
  studentCount: number;
  createdAt: string;
}

export type CreateGroupBodyTrainingType =
  (typeof CreateGroupBodyTrainingType)[keyof typeof CreateGroupBodyTrainingType];

export const CreateGroupBodyTrainingType = {
  online: "online",
  physical: "physical",
} as const;

export type CreateGroupBodyStatus =
  (typeof CreateGroupBodyStatus)[keyof typeof CreateGroupBodyStatus];

export const CreateGroupBodyStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface CreateGroupBody {
  name: string;
  startDate: string;
  trainingType: CreateGroupBodyTrainingType;
  capacity: number;
  notes?: string;
  status?: CreateGroupBodyStatus;
}

export type UpdateGroupBodyTrainingType =
  (typeof UpdateGroupBodyTrainingType)[keyof typeof UpdateGroupBodyTrainingType];

export const UpdateGroupBodyTrainingType = {
  online: "online",
  physical: "physical",
} as const;

export type UpdateGroupBodyStatus =
  (typeof UpdateGroupBodyStatus)[keyof typeof UpdateGroupBodyStatus];

export const UpdateGroupBodyStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface UpdateGroupBody {
  name?: string;
  startDate?: string;
  trainingType?: UpdateGroupBodyTrainingType;
  capacity?: number;
  /** @nullable */
  notes?: string | null;
  status?: UpdateGroupBodyStatus;
}

export type GroupWithStudentsTrainingType =
  (typeof GroupWithStudentsTrainingType)[keyof typeof GroupWithStudentsTrainingType];

export const GroupWithStudentsTrainingType = {
  online: "online",
  physical: "physical",
} as const;

export type GroupWithStudentsStatus =
  (typeof GroupWithStudentsStatus)[keyof typeof GroupWithStudentsStatus];

export const GroupWithStudentsStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface GroupWithStudents {
  id: number;
  name: string;
  startDate: string;
  trainingType: GroupWithStudentsTrainingType;
  capacity: number;
  /** @nullable */
  notes?: string | null;
  status: GroupWithStudentsStatus;
  studentCount: number;
  createdAt: string;
  students: Student[];
}

export interface ActivityLog {
  id: number;
  action: string;
  details: string;
  /** @nullable */
  performedBy?: string | null;
  createdAt: string;
}

export interface GalleryImage {
  id: number;
  url: string;
  /** @nullable */
  caption?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface UpdateGalleryImageBody {
  caption?: string;
  sortOrder?: number;
}

export interface DashboardStats {
  totalStudents: number;
  newStudents: number;
  contactedStudents: number;
  interestedStudents: number;
  noShowStudents: number;
  archivedStudents: number;
  totalGroups: number;
  openGroups: number;
}

export type ListStudentsParams = {
  stage?: string;
  trainingType?: string;
  search?: string;
  groupId?: number;
};

export type ListActivityParams = {
  limit?: number;
  offset?: number;
};

export type UploadGalleryImageBody = {
  image?: Blob;
  caption?: string;
  sortOrder?: number;
};