# lib/api-zod/src/generated/types/createGroupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAB SCHOOL CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGroupBodyStatus } from "./createGroupBodyStatus";
import type { CreateGroupBodyTrainingType } from "./createGroupBodyTrainingType";

export interface CreateGroupBody {
  name: string;
  startDate: Date;
  trainingType: CreateGroupBodyTrainingType;
  capacity: number;
  notes?: string;
  status?: CreateGroupBodyStatus;
}